Service folder created and naming corrections
